Chore: changed products page to products component
