+++ src/pages/Home/index.jsx
@@ -1,7 +1,9 @@
 import Searchbar from "components/Searchbar"
-import Products from "pages/Products"
+import Products from "components/Products"
 import Pagination from "components/Pagination"
 import Featured from "components/Featured"
+import useFetchProducts from "hooks/useFetchProducts"
+import {products} from "services"
 import {
   Container,
   Flex,
@@ -14,7 +16,7 @@ import "@fontsource/raleway"
 import "@fontsource/source-code-pro"
 
 const Home = () => {
-
+  const listOfProducts = useFetchProducts(products)
   return (
     <>
       <Container maxW="full" fontFamily="Gluten" color="teal" fontSize={80}>
@@ -32,7 +34,7 @@ const Home = () => {
             p={8}
             spacing={10}
             alignItems="center">
-            <Products v="100" h="20"/>
+            <Products v="100" h="20" items={listOfProducts}/>
             <Box alignContent="center" fontFamily="Source Code Pro">
               <Pagination/>
             </Box>
