--- src/pages/Products/index.jsx
@@ -1,26 +1,0 @@
-import { products } from "services";
-import { useEffect, useState } from "react";
-import ProductCard from "components/card";
-import { Flex } from "@chakra-ui/react";
-
-export default () => {
-  const [items, setItems] = useState([]);
-
-  useEffect(() => {
-    async function fetchData() {
-      const results = await products.getProducts(1, 5);
-      setItems((prevState) => [...prevState, ...results.products]);
-    }
-    fetchData();
-  }, []);
-
-  return (
-    <Flex d="flex">
-      {items.length ?
-      (items.map((item, i) => {
-          return <ProductCard key={i} {...item}/>;
-        })) :
-      (<h4>Loading...</h4>)}
-    </Flex>
-  );
-};
