+++ src/pages/index.js
@@ -1,4 +1,3 @@
-import Products from "pages/Products";
 import Home from "pages/Home";
 import SingleProduct from "pages/SingleProduct"
 
@@ -6,7 +5,6 @@ const App = () => {
 	return (
 		<>
 			<Home />
-			<Products />
 			<SingleProduct/>
 		</>
 	);
