+++ src/routes/index.js
@@ -1,13 +1,7 @@
-import Products from "pages/Products";
 import Home from "pages/Home";
 
 export default [
-  {
-    path: "/products",
-    component: Products,
-    exact: true,
-  },
-  {
+   {
     path: "/home",
     component: Home,
     exact: true,
